Add an integration test
