+++ src/main/java/be/ordina/ordineo/importusers/ScheduledTasks.java
@@ -1,31 +1,24 @@
 package be.ordina.ordineo.importusers;
 
-import be.ordina.ordineo.repository.EmployeeRepository;
 import be.ordina.ordineo.service.GitHubService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
-import java.text.SimpleDateFormat;
 
-/**
- * Created by SaFu on 19/04/2017.
- */
 @Component
 public class ScheduledTasks {
 
-    private static final Logger log = LoggerFactory.getLogger(ScheduledTasks.class);
-    private static final SimpleDateFormat dateFormat = new SimpleDateFormat("HH:mm:ss");
+    private final GitHubService gitHubService;
 
     @Autowired
-    private EmployeeRepository employeeRepository;
+    public ScheduledTasks(GitHubService gitHubService) {
+        this.gitHubService = gitHubService;
+    }
 
     @Scheduled(cron = "0 0 0 * * *")
     public void importGitHubUsers() throws IOException {
-        GitHubService gitHubService = new GitHubService(employeeRepository);
         gitHubService.importUsers();
     }
