--- src/test/resources/application-test.yml
@@ -1,6 +1,0 @@
-eureka:
-  client:
-    registerWithEureka: false
-    fetchRegistry: false
-server:
-  port: 8090
