+++ src/test/resources/application.yml
@@ -1,0 +1,9 @@
+eureka:
+  client:
+    enabled: false
+spring:
+  profiles:
+    active: test
+logging:
+  level:
+    org.springframework: ERROR
